Web: move code out into separate files
