+++ web/packages/teleport/src/Discover/SelectResource/SelectResource.tsx
@@ -16,49 +16,29 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  useEffect,
-  useMemo,
-  useState,
-  type ComponentPropsWithoutRef,
-} from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router';
 import styled from 'styled-components';
 
 import { Alert, Box, Flex, Link, P3, Text } from 'design';
 import * as Icons from 'design/Icon';
-import { NewTab } from 'design/Icon';
-import { getPlatform, Platform } from 'design/platform';
-import { Resource } from 'gen-proto-ts/teleport/userpreferences/v1/onboard_pb';
-import { UserPreferences } from 'gen-proto-ts/teleport/userpreferences/v1/userpreferences_pb';
+import { getPlatform } from 'design/platform';
 
 import AddApp from 'teleport/Apps/AddApp';
 import { FeatureHeader, FeatureHeaderTitle } from 'teleport/components/Layout';
-import { ToolTipNoPermBadge } from 'teleport/components/ToolTipNoPermBadge';
 import cfg from 'teleport/config';
-import {
-  BASE_RESOURCES,
-  getResourcePretitle,
-} from 'teleport/Discover/SelectResource/resources';
-import {
-  HeaderSubtitle,
-  PermissionsErrorMessage,
-  ResourceKind,
-} from 'teleport/Discover/Shared';
-import { resourceKindToPreferredResource } from 'teleport/Discover/Shared/ResourceKind';
+import { BASE_RESOURCES } from 'teleport/Discover/SelectResource/resources';
+import { HeaderSubtitle } from 'teleport/Discover/Shared';
 import { storageService } from 'teleport/services/storageService';
-import { Acl, AuthType, OnboardDiscover } from 'teleport/services/user';
 import { useUser } from 'teleport/User/UserContext';
 import useTeleport from 'teleport/useTeleport';
 
-import { getMarketingTermMatches } from './getMarketingTermMatches';
-import { DiscoverIcon } from './icons';
 import { SAML_APPLICATIONS } from './resources';
-import {
-  PrioritizedResources,
-  SearchResource,
-  type ResourceSpec,
-} from './types';
+import { Tile } from './Tile';
+import { SearchResource, type ResourceSpec } from './types';
+import { addHasAccessField } from './utils/checkAccess';
+import { filterBySupportedPlatformsAndAuthTypes } from './utils/filters';
+import { sortResourcesByKind, sortResourcesByPreferences } from './utils/sort';
 
 interface SelectResourceProps {
   onSelect: (resource: ResourceSpec) => void;
@@ -89,11 +69,11 @@ export function SelectResource({ onSelect }: SelectResourceProps) {
   const platform = getPlatform();
   const defaultResources: ResourceSpec[] = useMemo(
     () =>
-      sortResources(
+      sortResourcesByPreferences(
         // Apply access check to each resource.
         addHasAccessField(
           acl,
-          filterResources(
+          filterBySupportedPlatformsAndAuthTypes(
             platform,
             authType,
             getDefaultResources(cfg.isEnterprise)
@@ -193,98 +173,15 @@ export function SelectResource({ onSelect }: SelectResourceProps) {
       {resources && resources.length > 0 && (
         <>
           <Grid role="grid">
-            {resources.map((r, index) => {
-              const title = r.name;
-              const pretitle = getResourcePretitle(r);
-              const select = () => {
-                if (!r.hasAccess) {
-                  return;
-                }
-
-                setShowApp(true);
-                onSelect(r);
-              };
-
-              let resourceCardProps: ComponentPropsWithoutRef<
-                'button' | typeof Link
-              >;
-
-              if (r.kind === ResourceKind.Application && r.isDialog) {
-                resourceCardProps = {
-                  onClick: select,
-                  onKeyUp: (e: KeyboardEvent) => e.key === 'Enter' && select(),
-                  role: 'button',
-                };
-              } else if (r.unguidedLink) {
-                resourceCardProps = {
-                  as: Link,
-                  href: r.hasAccess ? r.unguidedLink : null,
-                  target: '_blank',
-                  style: { textDecoration: 'none' },
-                  role: 'link',
-                };
-              } else {
-                resourceCardProps = {
-                  onClick: () => r.hasAccess && onSelect(r),
-                  onKeyUp: (e: KeyboardEvent) => {
-                    if (e.key === 'Enter' && r.hasAccess) {
-                      onSelect(r);
-                    }
-                  },
-                  role: 'button',
-                };
-              }
-
-              // There can be three types of click behavior with the resource cards:
-              //  1) If the resource has no interactive UI flow ("unguided"),
-              //     clicking on the card will take a user to our docs page
-              //     on a new tab.
-              //  2) If the resource is guided, we start the "flow" by
-              //     taking user to the next step.
-              //  3) If the resource is kind 'Application', it will render the legacy
-              //     popup modal where it shows user to add app manually or automatically.
-              return (
-                <ResourceCard
-                  data-testid={r.kind}
-                  key={`${index}${pretitle}${title}`}
-                  hasAccess={r.hasAccess}
-                  aria-label={`${pretitle} ${title}`}
-                  {...resourceCardProps}
-                >
-                  {!r.unguidedLink && r.hasAccess && (
-                    <BadgeGuided>Guided</BadgeGuided>
-                  )}
-                  {!r.hasAccess && (
-                    <ToolTipNoPermBadge
-                      children={<PermissionsErrorMessage resource={r} />}
-                    />
-                  )}
-                  <Flex px={2} alignItems="center" height="48px">
-                    <Flex mr={3} justifyContent="center" width="24px">
-                      <DiscoverIcon name={r.icon} />
-                    </Flex>
-                    <Box>
-                      {pretitle && (
-                        <Text typography="body3" color="text.slightlyMuted">
-                          {pretitle}
-                        </Text>
-                      )}
-                      {r.unguidedLink ? (
-                        <Text bold color="text.main">
-                          {title}
-                        </Text>
-                      ) : (
-                        <Text bold>{title}</Text>
-                      )}
-                    </Box>
-                  </Flex>
-
-                  {r.unguidedLink && r.hasAccess ? (
-                    <NewTabInCorner color="text.muted" size={18} />
-                  ) : null}
-                </ResourceCard>
-              );
-            })}
+            {resources.map((r, index) => (
+              <Tile
+                // TODO(kimlisa): replace with r.id in upcoming PR
+                key={`${index}${r.name}${r.kind}`}
+                resourceSpec={r}
+                onChangeShowApp={setShowApp}
+                onSelectResource={onSelect}
+              />
+            ))}
           </Grid>
           <P3 mt={6}>
             Looking for something else?{' '}
@@ -338,358 +235,11 @@ const ClearSearch = ({ onClick }: { onClick(): void }) => {
   );
 };
 
-function checkHasAccess(acl: Acl, resourceKind: ResourceKind) {
-  const basePerm = acl.tokens.create;
-  if (!basePerm) {
-    return false;
-  }
-
-  switch (resourceKind) {
-    case ResourceKind.Application:
-      return acl.appServers.read && acl.appServers.list;
-    case ResourceKind.Database:
-      return acl.dbServers.read && acl.dbServers.list;
-    case ResourceKind.Desktop:
-      return acl.desktops.read && acl.desktops.list;
-    case ResourceKind.Kubernetes:
-      return acl.kubeServers.read && acl.kubeServers.list;
-    case ResourceKind.Server:
-      return acl.nodes.list;
-    case ResourceKind.SamlApplication:
-      return acl.samlIdpServiceProvider.create;
-    case ResourceKind.ConnectMyComputer:
-      // This is probably already true since without this permission the user wouldn't be able to
-      // add any other resource, but let's just leave it for completeness sake.
-      return acl.tokens.create;
-    default:
-      return false;
-  }
-}
-
-function sortResourcesByKind(
-  resourceKind: SearchResource,
-  resources: ResourceSpec[]
-) {
-  let sorted: ResourceSpec[] = [];
-  switch (resourceKind) {
-    case SearchResource.SERVER:
-      sorted = [
-        ...resources.filter(r => r.kind === ResourceKind.Server),
-        ...resources.filter(r => r.kind !== ResourceKind.Server),
-      ];
-      break;
-    case SearchResource.APPLICATION:
-      sorted = [
-        ...resources.filter(r => r.kind === ResourceKind.Application),
-        ...resources.filter(r => r.kind !== ResourceKind.Application),
-      ];
-      break;
-    case SearchResource.DATABASE:
-      sorted = [
-        ...resources.filter(r => r.kind === ResourceKind.Database),
-        ...resources.filter(r => r.kind !== ResourceKind.Database),
-      ];
-      break;
-    case SearchResource.DESKTOP:
-      sorted = [
-        ...resources.filter(r => r.kind === ResourceKind.Desktop),
-        ...resources.filter(r => r.kind !== ResourceKind.Desktop),
-      ];
-      break;
-    case SearchResource.KUBERNETES:
-      sorted = [
-        ...resources.filter(r => r.kind === ResourceKind.Kubernetes),
-        ...resources.filter(r => r.kind !== ResourceKind.Kubernetes),
-      ];
-      break;
-  }
-  return sorted;
-}
-
-const aBeforeB = -1;
-const aAfterB = 1;
-const aEqualsB = 0;
-
-/**
- * Evaluates the predicate and prioritizes the element matching the predicate over the element that
- * doesn't.
- *
- * @example
- * comparePredicate({color: 'green'}, {color: 'red'}, (el) => el.color === 'green') // => -1 (a before b)
- * comparePredicate({color: 'red'}, {color: 'green'}, (el) => el.color === 'green') // => 1  (a after  b)
- * comparePredicate({color: 'blue'}, {color: 'pink'}, (el) => el.color === 'green') // => 0  (both are equal)
- */
-function comparePredicate<ElementType>(
-  a: ElementType,
-  b: ElementType,
-  predicate: (resource: ElementType) => boolean
-): -1 | 0 | 1 {
-  const aMatches = predicate(a);
-  const bMatches = predicate(b);
-
-  if (aMatches && !bMatches) {
-    return aBeforeB;
-  }
-
-  if (bMatches && !aMatches) {
-    return aAfterB;
-  }
-
-  return aEqualsB;
-}
-
-export function sortResources(
-  resources: ResourceSpec[],
-  preferences: UserPreferences,
-  onboardDiscover: OnboardDiscover | undefined
-) {
-  const { preferredResources, hasPreferredResources } =
-    getPrioritizedResources(preferences);
-  const platform = getPlatform();
-
-  const sortedResources = [...resources];
-  const accessible = sortedResources.filter(r => r.hasAccess);
-  const restricted = sortedResources.filter(r => !r.hasAccess);
-
-  const hasNoResources = onboardDiscover && !onboardDiscover.hasResource;
-  const prefersServers =
-    hasPreferredResources &&
-    preferredResources.includes(
-      resourceKindToPreferredResource(ResourceKind.Server)
-    );
-  const prefersServersOrNoPreferences =
-    prefersServers || !hasPreferredResources;
-  const shouldShowConnectMyComputerFirst =
-    hasNoResources &&
-    prefersServersOrNoPreferences &&
-    isConnectMyComputerAvailable(accessible);
-
-  // Sort accessible resources by:
-  // 1. os
-  // 2. preferred
-  // 3. guided
-  // 4. alphabetically
-  //
-  // When available on the given platform, Connect My Computer is put either as the first resource
-  // if the user has no resources, otherwise it's at the end of the guided group.
-  accessible.sort((a, b) => {
-    const compareAB = (predicate: (r: ResourceSpec) => boolean) =>
-      comparePredicate(a, b, predicate);
-    const areBothGuided = !a.unguidedLink && !b.unguidedLink;
-
-    // Special cases for Connect My Computer.
-    // Show Connect My Computer tile as the first resource.
-    if (shouldShowConnectMyComputerFirst) {
-      const prioritizeConnectMyComputer = compareAB(
-        r => r.kind === ResourceKind.ConnectMyComputer
-      );
-      if (prioritizeConnectMyComputer) {
-        return prioritizeConnectMyComputer;
-      }
-
-      // Within the guided group, deprioritize server tiles of the current user platform if Connect
-      // My Computer is available.
-      //
-      // If the user has no resources available in the cluster, we want to nudge them towards
-      // Connect My Computer rather than, say, standalone macOS setup.
-      //
-      // Only do this if the user doesn't explicitly prefer servers. If they prefer servers, we
-      // want the servers for their platform to be displayed in their usual place so that the user
-      // doesn't miss that Teleport supports them.
-      if (!prefersServers && areBothGuided) {
-        const deprioritizeServerForUserPlatform = compareAB(
-          r => !(r.kind == ResourceKind.Server && r.platform === platform)
-        );
-        if (deprioritizeServerForUserPlatform) {
-          return deprioritizeServerForUserPlatform;
-        }
-      }
-    } else if (areBothGuided) {
-      // Show Connect My Computer tile as the last guided resource if the user already added some
-      // resources or they prefer other kinds of resources than servers.
-      const deprioritizeConnectMyComputer = compareAB(
-        r => r.kind !== ResourceKind.ConnectMyComputer
-      );
-      if (deprioritizeConnectMyComputer) {
-        return deprioritizeConnectMyComputer;
-      }
-    }
-
-    // Display platform resources first
-    const prioritizeUserPlatform = compareAB(r => r.platform === platform);
-    if (prioritizeUserPlatform) {
-      return prioritizeUserPlatform;
-    }
-
-    // Display preferred resources second
-    if (hasPreferredResources) {
-      const prioritizePreferredResource = compareAB(r =>
-        preferredResources.includes(resourceKindToPreferredResource(r.kind))
-      );
-      if (prioritizePreferredResource) {
-        return prioritizePreferredResource;
-      }
-    }
-
-    // Display guided resources third
-    const prioritizeGuided = compareAB(r => !r.unguidedLink);
-    if (prioritizeGuided) {
-      return prioritizeGuided;
-    }
-
-    // Alpha
-    return a.name.localeCompare(b.name);
-  });
-
-  // Sort restricted resources alphabetically
-  restricted.sort((a, b) => {
-    return a.name.localeCompare(b.name);
-  });
-
-  // Sort resources that user has access to the
-  // top of the list, so it is more visible to
-  // the user.
-  return [...accessible, ...restricted];
-}
-
-function isConnectMyComputerAvailable(
-  accessibleResources: ResourceSpec[]
-): boolean {
-  return !!accessibleResources.find(
-    resource => resource.kind === ResourceKind.ConnectMyComputer
-  );
-}
-
-/**
- * Returns prioritized resources based on user preferences cluster state
- *
- * @remarks
- * A user can have preferredResources set via onboarding either from the survey (preferredResources)
- * or various query parameters (marketingParams). We sort the list by the marketingParams if available.
- * If not, we sort by preferred resource type if available.
- * We do not search.
- *
- * @param preferences - Cluster state user preferences
- * @returns PrioritizedResources which is both the resource to prioritize and a boolean value of the value
- *
- */
-function getPrioritizedResources(
-  preferences: UserPreferences
-): PrioritizedResources {
-  const marketingParams = preferences.onboard.marketingParams;
-
-  if (marketingParams) {
-    const marketingPriorities = getMarketingTermMatches(marketingParams);
-    if (marketingPriorities.length > 0) {
-      return {
-        hasPreferredResources: true,
-        preferredResources: marketingPriorities,
-      };
-    }
-  }
-
-  const preferredResources = preferences.onboard.preferredResources || [];
-
-  // hasPreferredResources will be false if all resources are selected
-  const maxResources = Object.keys(Resource).length / 2 - 1;
-  const selectedAll = preferredResources.length === maxResources;
-
-  return {
-    preferredResources: preferredResources,
-    hasPreferredResources: preferredResources.length > 0 && !selectedAll,
-  };
-}
-
-export function filterResources(
-  platform: Platform,
-  authType: AuthType,
-  resources: ResourceSpec[]
-) {
-  return resources.filter(resource => {
-    const resourceSupportsPlatform =
-      !resource.supportedPlatforms?.length ||
-      resource.supportedPlatforms.includes(platform);
-
-    const resourceSupportsAuthType =
-      !resource.supportedAuthTypes?.length ||
-      resource.supportedAuthTypes.includes(authType);
-
-    return resourceSupportsPlatform && resourceSupportsAuthType;
-  });
-}
-
-function addHasAccessField(
-  acl: Acl,
-  resources: ResourceSpec[]
-): ResourceSpec[] {
-  return resources.map(r => {
-    const hasAccess = checkHasAccess(acl, r.kind);
-    switch (r.kind) {
-      case ResourceKind.Database:
-        return { ...r, dbMeta: { ...r.dbMeta }, hasAccess };
-      default:
-        return { ...r, hasAccess };
-    }
-  });
-}
-
 const Grid = styled.div`
   display: grid;
   grid-template-columns: repeat(auto-fill, 320px);
   column-gap: 10px;
   row-gap: 15px;
-`;
-
-const NewTabInCorner = styled(NewTab)`
-  position: absolute;
-  top: ${props => props.theme.space[3]}px;
-  right: ${props => props.theme.space[3]}px;
-  transition: color 0.3s;
-`;
-
-const ResourceCard = styled.button<{ hasAccess?: boolean }>`
-  position: relative;
-  text-align: left;
-  background: ${props => props.theme.colors.spotBackground[0]};
-  transition: all 0.3s;
-
-  border: none;
-  border-radius: 8px;
-  padding: 12px;
-  color: ${props => props.theme.colors.text.main};
-  line-height: inherit;
-  font-size: inherit;
-  font-family: inherit;
-  cursor: pointer;
-
-  opacity: ${props => (props.hasAccess ? '1' : '0.45')};
-
-  &:focus-visible {
-    outline: none;
-    box-shadow: 0 0 0 3px ${props => props.theme.colors.brand};
-  }
-
-  &:hover,
-  &:focus-visible {
-    background: ${props => props.theme.colors.spotBackground[1]};
-
-    ${NewTabInCorner} {
-      color: ${props => props.theme.colors.text.slightlyMuted};
-    }
-  }
-`;
-
-const BadgeGuided = styled.div`
-  position: absolute;
-  background: ${props => props.theme.colors.brand};
-  color: ${props => props.theme.colors.text.primaryInverse};
-  padding: 0px 6px;
-  border-top-right-radius: 8px;
-  border-bottom-left-radius: 8px;
-  top: 0px;
-  right: 0px;
-  font-size: 10px;
-  line-height: 24px;
 `;
 
 const InputWrapper = styled.div`
